refactor: renamed `consts` to `constants`
